split request classes into files
fixes #25
