+++ src/vbms.rb
@@ -10,3 +10,7 @@ require 'vbms/common'
 require 'vbms/client'
 require 'vbms/requests'
 require 'vbms/version'
+
+require 'vbms/requests/upload_document_with_associations'
+require 'vbms/requests/list_documents'
+require 'vbms/requests/fetch_document_by_id'
